+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JZSN3GD8S1WEAKMY5185CQFF"
+        "01JZW0NCDDTD6PHZTCDK6K8C82"
     )
     r"""The id of the resource"""