+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -59,5 +59,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-07-10T07:51:45.591Z")
+    created: Optional[datetime] = parse_datetime("2025-07-11T05:52:14.598Z")
     r"""The date and time the resource was created"""