+++ packages/orq-rc/RELEASES.md
@@ -1949,3 +1949,13 @@ Based on:
 - [python v3.10.0-rc.3] packages/orq-rc
 ### Releases
 - [PyPI v3.10.0-rc.3] https://pypi.org/project/orq-ai-sdk/3.10.0-rc.3 - packages/orq-rc
+
+## 2025-07-11 05:53:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.580.2 (2.656.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.10.0-rc.4] packages/orq-rc
+### Releases
+- [PyPI v3.10.0-rc.4] https://pypi.org/project/orq-ai-sdk/3.10.0-rc.4 - packages/orq-rc